requirements.txt
langchain>=0.1.17
langchain-openai>=0.0.7
langchain-community>=0.0.11
python-dotenv
flask
pyyaml
openai